Fixed const/non-const access for Nodes and Cells.
All Topologies and Geometries (except the now-defunct RectilinearMesh)
now supply full node and cell access. For the case of structured meshes,
the nodes supply normal referencing while the connectivity array is
independent of the true geometry (since connectivity cannot be changed
for a structured mesh). In this case, the connectivity array is lazily
constructed only if needed.